hadoop: Implement StreamCapabilities
Introduced the StreamCapabilities API and the apabilities `UNBUFFER`,
`READBYTEBUFFER`, and `READHEAD`. This streamlines use by many common
tools (such as spark, parquet) which typically interact with HDFS via
the `ByteBufferReadable` interface when reading, but fall back to
copying the data into an array (when directly allocated) or, potentially
unboxing it in the best cases, just for `QFSInputStream` to then
subsequently wrap the given array in a buffer.

Additionally for reading text files using the HDFS API, do not allocate
a single byte array to wrap to invoke the `read([]byte,int,int)` method which
then wraps the array in a byte buffer. Replaced with exposing a `read()`
operation on `KfsInputChannel` which follows the same protocol, simply
unbufferring a single byte at a time from the `KfsInputChannel` buffer.

This also fixes a very silly bug where byte-by-byte reads were double
counted in the filesystem statistics.

Fixed the implementation of `setReadahead()` which was incorrectly
labeled `setReadaheadSize()` which was never an HDFS operation.

Implemented the respective interfaces in `QFSInputStream`,
`CanUnbuffer`, `CanSetReadahead` in addition to the aforementioned.